requirements.txt
asgiref==3.2.10 
django==3.1.1
pytz==2020.1 
sqlparse==0.3.1
djangorestframework==3.12.1
PyJWT==1.7.1